fix: fix model param mapping
